Added Profile Update
when user goes to /account/profile their data are loaded in the form. They can edit those data and submit form.
profileupdate in view will handle the request and update the details and redirect to /account/profile showing flash message.
address field along with other fields depends on other models which are not complete hence they are not updated.